add view composers for menu and language, implement storage link command
